Move boardState
