+++ front/src/modules/pipeline-progress/states/boardState.ts
