+++ front/src/modules/pipeline-progress/states/currentPipelineState.ts
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+import { Pipeline } from '~/generated/graphql';
+
+export const currentPipelineState = atom<Pipeline | undefined>({
+  key: 'currentPipelineState',
+  default: undefined,
+});
