+++ front/src/modules/pipeline-progress/states/isBoardLoadedState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const isBoardLoadedState = atom<boolean>({
+  key: 'isBoardLoadedState',
+  default: false,
+});
